This commit fixes some issues I made while merging. Mainly that parts of the permabrig were considered to be space. That'll teach me to see only selectable areas...
ALSO, this adapts the BS12 police tape code for our server, and I made new icons for the actual taperoll item, because the old ones were shit. Police taperolls were added to sec lockers in security, (not the posts), and engi tape was added to crates inside atmos and engineering. Copying and pasting my explanation from the thread:

It includes both police and engineering tape, (Brig access is needed to pass through police tape, and engineering or atmos access is needed for engi tape) with new icons I made, because the old ones were shit. Tape works like this:

-Click on tape in hand and walk vertical or horizontal.
-When you've reached where you want to go, click it again.

-Police tape has a special feature, in which you can click on doors to tape them up.

-To break tape, click on it with an empty hand.

-To break tape if you have the proper ID, click it without help intent on, clicking it with help intent will raise it and allow people without the proper ID to pass through.