Map changes, yo
A few additions and changes to the map.

- Changed some rwalls near the armory to windows, to allow another way
of breaking in.
- Added Nien's energy pistol to the HoS' desk, the alternate RD's outfit
to his dorm, and the CMO's stamp in their office. Replica katana (5
brute) in maintenance.
- Put windows to the west of the AI upload, giving another
halfway-decent point of entry.
- Replaced the fitness room toilets with the art storage, added a secret
torture room where the art storage used to be.
- Added more extinguisher cabinets in various places. Chemistry,
security, civilian sector... Fire-safety closets, too.
- Pair of pump shotguns added to armory. They're loaded with beanbag
shells.
- Removed random piece of cable under medbay maintenance door, by the
morgue.
- Shifts the floor of R&D over one tile, so you can use all four
machines at once. Added a table with science goggles to make it seem
less empty.
- Shuffled the crates in cargo maintenance, hid a Russian surplus locker
between the warehouse and sec.
Pictures of the new spots. http://imgur.com/a/aBKxZ#0