Fixes mail system
After a long session of staring at pipes, right-click-> delete-ing, and
editing variables, I finally managed to fix the mail system. I added
disposals to every place on the map where there should be one and fixed
the sorters for all the other ones. I had to gut large parts of the
disposals near medbay, but I did it.